Extracted common key file methods into an abstract base class
